Revert to header file
